Assigns package-scoped script_name.
By default this is sourced from the focused RStudio window or the calling command (e.g. Rscript script_name.R)

Assigns package-scoped script_name.
By default this is sourced from the focused RStudio window or the calling command (e.g. Rscript script_name.R)

Usage

set_script_name(script_name = "")

Arguments

script_name

optional arg to override the calling script